Make option for adding kmod to extraModulePackages
This allows you to keep building your system kernel with the needed
KALLSYMS_ALL config option but otherwise disable building the memflow
kernel module itself. This is useful if the kernel module ever fails to
build but you wish to renable it in the future without having to fully
rebuild your kernel again.